Fixing submission 1 dan menambahkan maps activity

+++ app/src/main/java/com/andricohalim/storyapp/response/StoryResponse.kt
@@ -34,8 +34,8 @@ data class ListStoryItem(
 	val description: String,
 
 	@field:SerializedName("lon")
-	val lon: Double? = null,
+	val lon: Float? = null,
 
 	@field:SerializedName("lat")
-	val lat: Double? = null
+	val lat: Float? = null
 ) : Parcelable
